(1248) Rename and normalize shantytownCommentModel
