+++ packages/api/server/models/shantytownCommentModel/index.js
@@ -2,8 +2,8 @@ const create = require('./create');
 const findOne = require('./findOne');
 const getHistory = require('./getHistory');
 
-module.exports = {
+module.exports = () => ({
     create,
     findOne,
     getHistory,
-};
+});
