+++ packages/api/test/suites/server/services/shantytownComment/createComment.spec.js
@@ -7,7 +7,7 @@ chai.use(sinonChai);
 
 const { expect } = chai;
 const { sequelize } = require('#db/models');
-const shantytownCommentModel = require('#server/models/shantytownComment');
+const shantytownCommentModel = require('#server/models/shantytownCommentModel')();
 const shantytownModel = require('#server/models/shantytownModel')(sequelize);
 const mattermostUtils = require('#server/utils/mattermost');
 const userModel = require('#server/models/userModel')(sequelize);
